To Determine if a person is diabetic or not
Use case - 
to Determine if a patinet is diabetic or not based on certain features like

Pregnancies – The number of times a woman has been pregnant.
Glucose – The blood sugar level in the body, which is important for diagnosing diabetes.
BloodPressure – The pressure of blood against the walls of the arteries, indicating heart health.
SkinThickness – The thickness of skin measured in millimeters, often used to estimate body fat.
Insulin – The amount of insulin (a hormone that regulates blood sugar) present in the blood.
BMI (Body Mass Index) – A measure of body fat based on height and weight.
DiabetesPedigreeFunction – A score that estimates the likelihood of diabetes based on family history and genetic factors.

It is Supervised ML